[checked-build] Assert if we safepoint while an OS mutex is locked
Suppose two threads both try to lock an OS mutex while in GC Unsafe mode.  If
one of them locks the mutex and then hits a safepoint and suspends, the other
thread will block and wait for the mutex while in GC Unsafe (cooperative) mode.
As a result, the suspend initiator (the GC) will deadlock and wait for the
thread forever.

This PR adds a "no safepoints" state machine transition in checked builds (set
MONO_CHECK_MODE=thread) that will cause an assertion if a thread attempt to
safepoint while holding an OS mutex.