[utils] Implement MonoCoopBarrier
A barrier is a synchronization primitive that lets some number of participant
threads to wait until all of them reach the barrier and the proceed in unison.

This barrier implementation is copied from the reference source
System.Threading.Barrier implementation, except without support for a
cancellation token.

The reason we need our own instead of (for example) using the POSIX one is
because we want to be able to add and remove participant threads after the
barrier is created.

This version is coop-aware in the sense that mono_coop_barrier_wait and
mono_coop_barrier_timedwait will put the thread in GC Safe mode while blocked.